requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-autofixture==0.12.1
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-like-systems==0.0.1
django-utils-six==2.0
idna==2.10
Pillow==8.1.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
urllib3==1.26.3
